ipwe

+++ ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: agrodzin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/22 09:58:44 by agrodzin          #+#    #+#             */
-/*   Updated: 2018/03/21 18:15:39 by agrodzin         ###   ########.fr       */
+/*   Updated: 2018/03/21 18:17:24 by agrodzin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned char	*ss1;
 	unsigned char	*ss2;
